+++ packages/plugins/typescript/graphql-request/tests/test-files/run-example-query.ts
@@ -1,0 +1,8 @@
+import { GraphQLClient } from 'graphql-request';
+import { getSdk } from './graphql-request-sdk';
+
+export function runExampleQuery(x: number, y: number) {
+  const client = new GraphQLClient('http://localhost:4000/graphql');
+  const sdk = getSdk(client);
+  return sdk.Add({ x, y });
+}