+++ packages/plugins/typescript/react-query/src/fetcher-graphql-request.ts
@@ -22,7 +22,9 @@ function fetcher<TData, TVariables>(client: GraphQLClient, query: string, variab
     hasRequiredVariables: boolean
   ): string {
     const variables = generateQueryVariablesSignature(hasRequiredVariables, operationVariablesTypes);
-    this.visitor.imports.add(`import { GraphQLClient } from 'graphql-request';`);
+
+    const typeImport = this.visitor.config.useTypeImports ? 'import type' : 'import';
+    this.visitor.imports.add(`${typeImport} { GraphQLClient } from 'graphql-request';`);
 
     const hookConfig = this.visitor.queryMethodMap;
     this.visitor.reactQueryIdentifiersInUse.add(hookConfig.query.hook);