+++ packages/plugins/typescript/react-query/tests/react-query.spec.ts
@@ -44,12 +44,32 @@ describe('React-Query', () => {
       }
     }
   `);
+  const basicFragment = parse(/* GraphQL */ `
+    fragment EntryData on Entry {
+      feed {
+        id
+        commentCount
+        repository {
+          full_name
+          html_url
+          owner {
+            avatar_url
+          }
+        }
+      }
+    }
+  `);
   const docs = [
     {
       document: basicDoc,
     },
     {
       document: basicMutation,
+    },
+  ];
+  const notOperationDocs = [
+    {
+      document: basicFragment,
     },
   ];
 
@@ -137,6 +157,46 @@ describe('React-Query', () => {
       expect(out.content).toMatchSnapshot();
       await validateTypeScript(mergeOutputs(out), schema, docs, config, false);
     });
+
+    it('Should generate mutation correctly with lazy variables', async () => {
+      const config = {
+        fetcher: {
+          func: './my-file#useCustomFetcher',
+          isReactHook: true,
+        },
+        typesPrefix: 'T',
+      };
+
+      const out = (await plugin(schema, docs, config)) as Types.ComplexPluginOutput;
+
+      expect(out.prepend).toContain(
+        `import { useQuery, UseQueryOptions, useMutation, UseMutationOptions } from 'react-query';`
+      );
+      expect(out.prepend).toContain(`import { useCustomFetcher } from './my-file';`);
+      expect(out.content).toBeSimilarStringTo(`export const useTestQuery = <
+          TData = TTestQuery,
+          TError = unknown
+        >(
+          variables?: TTestQueryVariables, 
+          options?: UseQueryOptions<TTestQuery, TError, TData>
+        ) => 
+        useQuery<TTestQuery, TError, TData>(
+          ['test', variables],
+          useCustomFetcher<TTestQuery, TTestQueryVariables>(TestDocument).bind(null, variables),
+          options
+        );`);
+      expect(out.content).toBeSimilarStringTo(`export const useTestMutation = <
+        TError = unknown,
+        TContext = unknown
+      >(options?: UseMutationOptions<TTestMutation, TError, TTestMutationVariables, TContext>) => 
+      useMutation<TTestMutation, TError, TTestMutationVariables, TContext>(
+        useCustomFetcher<TTestMutation, TTestMutationVariables>(TestDocument),
+        options
+      );`);
+
+      expect(out.content).toMatchSnapshot();
+      await validateTypeScript(mergeOutputs(out), schema, docs, config, false);
+    });
   });
 
   describe('fetcher: graphql-request', () => {
@@ -183,6 +243,16 @@ describe('React-Query', () => {
 
       expect(out.content).toMatchSnapshot();
       await validateTypeScript(mergeOutputs(out), schema, docs, config, false);
+    });
+    it('Should support useTypeImports', async () => {
+      const config = {
+        fetcher: 'graphql-request',
+        useTypeImports: true,
+      };
+
+      const out = (await plugin(schema, docs, config)) as Types.ComplexPluginOutput;
+
+      expect(out.prepend).toContain(`import type { GraphQLClient } from 'graphql-request';`);
     });
   });
 
@@ -409,5 +479,17 @@ describe('React-Query', () => {
         `useTestQuery.getKey = (variables?: TestQueryVariables) => ['test', variables];`
       );
     });
+  });
+
+  it('Should not generate fetcher if there are no operations', async () => {
+    const out = (await plugin(schema, notOperationDocs, {})) as Types.ComplexPluginOutput;
+    expect(out.prepend).not.toBeSimilarStringTo(`function fetcher<TData, TVariables>(`);
+
+    const config = {
+      fetcher: 'graphql-request',
+    };
+
+    const outGraphqlRequest = (await plugin(schema, notOperationDocs, config)) as Types.ComplexPluginOutput;
+    expect(outGraphqlRequest.prepend).not.toContain(`import { GraphQLClient } from 'graphql-request';`);
   });
 });