+++ packages/plugins/typescript/typed-document-node/src/index.ts
@@ -2,7 +2,12 @@ import { Types, PluginValidateFn, PluginFunction } from '@graphql-codegen/plugin
 import { visit, concatAST, GraphQLSchema, Kind, FragmentDefinitionNode } from 'graphql';
 import { TypeScriptTypedDocumentNodesConfig } from 'packages/plugins/typescript/typed-document-node/src/config';
 import { extname } from 'path';
-import { LoadedFragment, RawClientSideBasePluginConfig, DocumentMode, optimizeOperations } from '@graphql-codegen/visitor-plugin-common';
+import {
+  LoadedFragment,
+  RawClientSideBasePluginConfig,
+  DocumentMode,
+  optimizeOperations,
+} from '@graphql-codegen/visitor-plugin-common';
 import { TypeScriptDocumentNodesVisitor } from './visitor';
 
 export const plugin: PluginFunction<TypeScriptTypedDocumentNodesConfig> = (