+++ packages/plugins/typescript/react-query/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @graphql-codegen/typescript-react-query
 
+## 1.2.2
+
+### Patch Changes
+
+- e69869aa: enhance(react-query): Don't generate fetcher without operations (#5601)
+
+## 1.2.1
+
+### Patch Changes
+
+- 70628c87: Add support for useTypeImports when using fetcher: graphql-request
+
+## 1.2.0
+
+### Minor Changes
+
+- ed0ae4c7: Allow fetcher to receive variables lazily so it can use react hooks
+
+### Patch Changes
+
+- Updated dependencies [5749cb8a]
+- Updated dependencies [5a12fe58]
+  - @graphql-codegen/visitor-plugin-common@1.18.3
+
 ## 1.1.0
 
 ### Minor Changes